# docs/swagger.yaml
definitions:
  domain.Task:
    description: A task entity with its properties
    properties:
      created_at:
        description: The timestamp when the task was created
        example: "2024-02-27T12:00:00Z"
        type: string
      deleted_at:
        description: The timestamp when the task was deleted (soft delete)
        example: "2024-02-27T12:00:00Z"
        format: date-time
        type: string
      description:
        description: A detailed description of the task
        example: Write comprehensive documentation for the API endpoints
        type: string
      id:
        description: The unique identifier of the task
        example: 1
        type: integer
      status:
        description: The current status of the task (pending, in_progress, completed)
        example: pending
        type: string
      title:
        description: The title of the task
        example: Complete project documentation
        type: string
      updated_at:
        description: The timestamp when the task was last updated
        example: "2024-02-27T12:00:00Z"
        type: string
    type: object
  healthcheck.Status:
    properties:
      messages:
        items:
          type: string
        type: array
      name:
        type: string
      up:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/tasks:
    get:
      description: Retrieve a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the provided details
      parameters:
      - description: Task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new task
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid task ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Retrieve a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Invalid task ID
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update a task with the provided details
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Invalid task ID or bad request
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a task
      tags:
      - tasks
  /healthcheck:
    get:
      operationId: v1-healthcheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthcheck.Status'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/healthcheck.Status'
      summary: Allows healthcheck of server in general and substatuses of various
        internal services
      tags:
      - test
swagger: "2.0"